pyproject.toml
[tool.poetry]
name = "uptime-robot"
version = "0.1.0"
description = ""
authors = ["AlexanderHura <alexander_gura@ukr.net>"]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.8.10"
django = "3.2"
django-environ = "^0.10.0"
psycopg2-binary = "==2.9.3"
celery = "==5.2.3"
django-celery-beat = "==2.2.1"



[tool.poetry.group.dev.dependencies]
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
